Complete ADS-1

+++ src/alg.cpp
@@ -1,12 +1,13 @@
 // Copyright 2022 NNTU-CS
 #include <cstdint>
-#include "alg.h"
 #include <math.h>
+#include "alg.h"
+
 
 
 bool checkPrime(uint64_t value) {
-    for(int i = 2;i <= sqrt(value);i++){
-        if(value % i == 0){
+    for (int i = 2;i <= sqrt(value);i++) {
+        if (value % i == 0) {
             return false;
         }
     }
@@ -15,9 +16,9 @@ bool checkPrime(uint64_t value) {
 
 uint64_t nPrime(uint64_t n) {
     int f = 1;
-    for (uint64_t i = 0;i < n;i++){
-        for(uint64_t j = f+1;j < (pow(2,64));j++){
-            if(checkPrime(j)){
+    for (uint64_t i = 0; i < n; i++) {
+        for (uint64_t j = f+1; j < (pow(2,64)); j++) {
+            if (checkPrime(j)) {
                 f = j;
                 break;
             }
@@ -28,7 +29,7 @@ uint64_t nPrime(uint64_t n) {
 
 uint64_t nextPrime(uint64_t value) {
     for (uint64_t j = value + 1; j < (pow(2, 64)); j++) {
-        if(checkPrime(j)){
+        if (checkPrime(j)) {
             value = j;
             break;
         }
@@ -38,9 +39,9 @@ uint64_t nextPrime(uint64_t value) {
 
 
 uint64_t sumPrime(uint64_t hbound) {
-    int sum = 0;
+    uint64_t sum = 0;
     for (uint64_t i = 2; i < hbound; ++i) {
-        if(checkPrime(i)){
+        if (checkPrime(i)) {
             sum += i;
         }
     }
